Fixed slightly embarrassing bug in timestamps - was explicitly casting long longs to ints. Once again I'm surprised this even worked. Probably wouldn't have worked on a big-endian system, or indeed on any system for very long. *updates test plan*

+++ trad4/objects/trad4.c
@@ -277,7 +277,7 @@ void set_timestamp( obj_loc_t obj_loc, int id )
 
     //std::cout << "setting timestamp: " << timestamp << std::endl;
 
-    *(int*)obj_loc[id] = timestamp;
+    *(long long*)obj_loc[id] = timestamp;
 
     ((object_header*)obj_loc[id])->status = STOPPED;
 }
